Replace popstation.c with popstation_toc.c
